Add files via upload
These are current versions of the C code I use for simulations on my YouTube page: https://www.youtube.com/c/NilsBerglund/
Instructions for compiling the code and generating a movie are in the comments at the beginning of the files. Please be aware that the code is still under development, and may contain bugs.